constrain_norm and get_fisher_inverse_fn functions moved to sr.py, proxsr.py (unify later)
